+++ src/main/resources/db/stockExchange.sql
@@ -13,7 +13,7 @@ DROP TABLE IF EXISTS user;
 CREATE TABLE IF NOT EXISTS user (
     user_id bigint NOT NULL AUTO_INCREMENT COMMENT '회원 아이디',
     user_name varchar(20) NOT NULL COMMENT '회원명',
-    user_krw_price bigint NOT NULL COMMENT '보유 원화 현금',
+    user_krw_price decimal(20, 2) NOT NULL COMMENT '보유 원화 현금',
     created_at timestamp NOT NULL COMMENT '가입 시간',
     PRIMARY KEY (user_id)
 ) ENGINE=INNODB COMMENT='회원';
@@ -24,7 +24,7 @@ CREATE TABLE IF NOT EXISTS stock (
     stock_name varchar(255) NOT NULL COMMENT '종목명',
     stock_number varchar(10) NOT NULL COMMENT '종목 코드',
     stock_ipo enum('KOSPI','KOSDAQ') NOT NULL COMMENT '상장시장',
-    stock_price int NOT NULL COMMENT '종목 가격',
+    stock_price decimal(12, 2) NOT NULL COMMENT '종목 가격',
     PRIMARY KEY (stock_id)
 ) ENGINE=INNODB COMMENT='주식 종목';
 
@@ -32,7 +32,7 @@ CREATE TABLE IF NOT EXISTS stock (
 CREATE TABLE IF NOT EXISTS order_table (
     order_id bigint NOT NULL AUTO_INCREMENT COMMENT '주문 아이디',
     order_count int NOT NULL COMMENT '주문 수량',
-    order_price bigint NOT NULL COMMENT '주문 가격',
+    order_price decimal(12, 2) NOT NULL COMMENT '주문 가격',
     order_type enum('BUY','SELL') NOT NULL COMMENT '주문 타입',
     order_status enum('PENDING','COMPLETED','CANCELLED') NOT NULL COMMENT '주문 상태',
     order_remain_count int NOT NULL COMMENT '미체결 수량',
@@ -51,7 +51,7 @@ CREATE TABLE IF NOT EXISTS order_table (
 CREATE TABLE IF NOT EXISTS execution (
     execution_id bigint NOT NULL AUTO_INCREMENT COMMENT '체결 아이디',
     execution_count int NOT NULL COMMENT '체결 수량',
-    execution_price bigint NOT NULL COMMENT '체결 가격',
+    execution_price decimal(12, 2) NOT NULL COMMENT '체결 가격',
     created_at timestamp NOT NULL COMMENT '체결 시간',
     execution_buy_order_id bigint NULL COMMENT '매수 주문 체결', -- Null 허용
     execution_sell_order_id bigint NULL COMMENT '매도 주문 체결', -- Null 허용
@@ -66,7 +66,7 @@ CREATE TABLE IF NOT EXISTS execution (
 CREATE TABLE IF NOT EXISTS holding (
     holding_id bigint NOT NULL AUTO_INCREMENT COMMENT '보유주식 아이디',
     holding_quantity int NOT NULL COMMENT '보유 수량',
-    holding_total_price bigint NOT NULL COMMENT '총 손익',
+    holding_total_price decimal(20, 2) NOT NULL COMMENT '총 손익',
     updated_at timestamp NOT NULL COMMENT '수정 시간',
     user_id bigint NOT NULL COMMENT '회원 아이디',
     stock_id bigint NOT NULL COMMENT '주식 아이디',
